Added versioning information and distribution support.
The game signals its version both in editor mode as well as in
play mode during the menu state.
The first log line now contains the version and commit timestamp.
The current build's version is also displayed in the editor and
a mechanism to check for new available versions and notify the user
is put in place.
sd.sh is updated to provide update-version capabilities.