When changing brains, attributes seem to be lost #6
Fixed by specifying a key comparator when defining attribute bag maps.
This was a terrible example of sloppy c++ programming. Java seems to be
cluttering my mind :)
I chose to use a custom comparator rather than using std::string as key
from a performance point of view. This could be called premature optimization
especially since it wasn't really measured but more of a educated guess.
This fixes and closes #6.