Level editor: expose entity weapons for editing #33
Decided to go on a different route than the initial suggestion in the issue description.
Removed weapon editing from the editor and instead added a new Lua abstration, Weapon.
Bullet emitters can be added to weapons and weapons instead can be attached to entities.
Updated manual to reflect the changes.
This fixes #33.